+++ packages/typedoc/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @to-skills/typedoc
 
+## 1.0.8
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @to-skills/core@1.3.0
+
 ## 1.0.7
 
 ### Patch Changes