+++ packages/vitepress/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @to-skills/vitepress
 
+## 0.2.25
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @to-skills/core@1.3.0
+
 ## 0.2.24
 
 ### Patch Changes